Add files via upload
The goal of the extended overhaul is to modularize the code and to facilitate new functionality by easy swapping of networks and data. Currently functionality is still restricted and namespaces have been broken, but the networks still run. Additionally the analysis part has been separated into a new bit of code aNEUlysis.